+++ cg3/data_structures/trees/includes/rangetree_helpers.tpp
@@ -193,7 +193,7 @@ inline void rebalanceRangeTreeHelper(
 
     //Not balanced node
     Node* n = node;
-    int balanceFactor = getHeightHelper(n->right) - getHeightHelper(n->left);
+    int balanceFactor = (int)(getHeightHelper(n->right) - getHeightHelper(n->left));
 
     //Climb on parents to find the not balanced node
     while (n != nullptr && balanceFactor >= -1 && balanceFactor <= 1) {
@@ -201,7 +201,7 @@ inline void rebalanceRangeTreeHelper(
 
         if (n != nullptr) {
             //Compute balance factor
-            balanceFactor = getHeightHelper(n->right) - getHeightHelper(n->left);
+            balanceFactor = (int)(getHeightHelper(n->right) - getHeightHelper(n->left));
 
             assert(balanceFactor <= 2 && balanceFactor >= -2);
         }
